rendered data into input fields, beginning css
