+++ septa-fare-calculator/septa-fare-calculator/src/components/SeptaFareCalculator/calculator.css
@@ -1,0 +1,9 @@
+.calculator-container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+
+    width: 400px;
+    border: 2px gray solid;
+}
